refactor: make the tool surface a library any host can register
The 34 tools now have two hosts: this extension, and the remote server that
pins this repository as a dependency. Nothing about them was host-agnostic —
sql.js was called directly from catalog/pages/services, and the wasm binary
came in through an import only esbuild can resolve, which forced a stub plugin
on the test runner and would have forced a WebAssembly build of SQLite on a
consumer that already has a native driver.

Three seams, one source tree:

- `ShamelaDb` (src/server/db.ts) is now the only way anything here opens
  SQLite — master.db, the per-book files behind the LRU-of-50, the service
  indexes. `sqljs.ts` implements it for the extension and takes the wasm bytes
  rather than importing them.
- `Helper` is an interface; the subprocess implementation is `JavaHelper`.
- `registerAllTools(server, deps)` registers everything on a server the caller
  owns. `deps` is the whole of what differs between hosts: where the install
  is, how to read its SQLite, how to reach the search engine.

`src/server/entry.ts` is the extension, and the first consumer of that API —
one code path, not two. It holds the `.wasm` import and nothing else does, so
the vitest wasm-stub plugin is gone.

Packaging: `dist/index.js` still comes from esbuild for the `.mcpb`, unchanged;
`dist/lib/` now comes from tsc for consumers, with an `exports` map and a
`prepare` script so a git install builds it. `zod` and the MCP SDK became peer
dependencies, and zod stays on 3 — the SDK's zod-4 branch emits input schemas
without `additionalProperties: false`, which would change all 34 tools' declared
shape. Measured, and written up in docs/decisions.md §١٣.

No tool added, removed, renamed or reshaped: 911 tests pass, the packed bundle
still advertises the same 34 tools over stdio with the same schemas.
