+++ docs/decisions.md
@@ -6,6 +6,47 @@
 
 ---
 
+## ١٣) شكلُ المخطَّطات ملكُ هذا المستودع، و`zod` تبقى على الإصدار الثالث
+
+**السياق.** صار هذا المستودع مكتبةً يستهلكها خادم الاستضافة (`shamela-mcp-server`)، والخادم يملك نسختَه من `@modelcontextprotocol/sdk` ومن `zod`. وكان عنده `zod@4` وهنا `zod@^3.24.1`. فالسؤال الذي لا مفرَّ منه: مَن يملك شكل المدخلات المعلَن للعميل؟
+
+**ما كشفه القياس.** الـ SDK يقبل الإصدارين معًا، لكنه **لا يحوّلهما بمحوِّلٍ واحد**: شكلُ زود ٣ يمرّ على `zod-to-json-schema`، وشكلُ زود ٤ على `toJSONSchema` المدمج في زود نفسها (`server/zod-json-schema-compat.js`). والفرق ليس تجميليًّا. قِيس على شكلٍ واحد بالمحوِّلَين، والخارج من الأداة نفسها:
+
+```
+zod 3 → { "type":"object", "properties":{…}, "required":["query"], "additionalProperties": false }
+zod 4 → { "type":"object", "properties":{…}, "required":["query"] }
+```
+
+فترقيةُ زود هنا تحذف `additionalProperties: false` من مخطَّط **كلِّ** أداةٍ من الأربع والثلاثين: عميلٌ كان يرفض حقلًا غير معلَن يصير يقبله. وهذا تغيُّرٌ في الواجهة المعلَنة، لا إعادةُ ترتيبٍ داخلية.
+
+**وقيسَ ما لا يتغيّر أيضًا.** الـ SDK يفحص **الشكل** لا الإصدارَ المثبَّت (`isZ4Schema`): فأشكالُ زود ٣٫٢٥٫٧٦ داخلَ مضيفٍ عنده زود ٤٫٤٫٣ تعمل، ويخرج المخطَّط نفسه بحذافيره، وينجح النداء. فالخطر ليس في تجاور الإصدارين، بل في أن تُبنى أشكالُنا بالإصدار الرابع.
+
+**القرار.** الأشكال تبقى زود ٣، ويُعلَن ذلك عقدًا لا عرَضًا:
+
+- `zod` و`@modelcontextprotocol/sdk` صارتا `peerDependencies` (ومعهما `devDependencies` ليبقى بناءُ هذا المستودع مكتفيًا بنفسه).
+- ورُفِع الحدُّ الأدنى لزود من `^3.24.1` إلى `^3.25.76`، لأن `zod-compat` في الـ SDK يستورد `zod/v3` و`zod/v4-mini`، وهما مساران لا وجود لهما قبل ٣٫٢٥ — أي أن الحدَّ القديم كان يجيز تركيبًا لا يُقلع أصلًا.
+
+ولماذا نظيرتان (`peer`) لا تبعيّتان؟ لأن `McpServer` **نفسه** يعبر الحدّ: المضيف ينشئه ونحن نسجّل عليه. ونسختان من الـ SDK تعنيان صنفين مختلفَين لكلٍّ منهما أعضاؤه الخاصّة، وTypeScript ترفض التوفيق بينهما. وأمّا زود فلأنها صارت نظيرة، فإصدارُ المضيف هو الذي **تُبنى به أشكالُنا** — ولو كان الرابع لعاد الفرقُ المقيس أعلاه من البابِ الذي أُغلق.
+
+**ثمنه.** خادمُ الاستضافة ينزل بزود إلى الإصدار الثالث. و`npm` يرفض غير ذلك رفضًا صريحًا يسمّي القيد (`peer zod@"^3.25.76" from shamela-mcp`)، ولا يُتجاوز إلا بـ`--force`. وهذا هو المقصود: خطأُ تثبيتٍ يُقرأ، خيرٌ من أربعةٍ وثلاثين مخطَّطًا تتغيّر في صمت. ويبقى البابُ إلى زود ٤ مفتوحًا يومَ يُقصد قصدًا: تُرفع الأشكال، ويُعاد ضبط لقطة `tests/snapshots/tool-schemas.json`، ويُعلَن التغيير في إصدارٍ كبير — لا أن يقع تبعًا لترقيةٍ في مستودعٍ آخر.
+
+---
+
+## ١٢) الأدواتُ لا تعرف من يستضيفها: واجهةٌ للـ SQLite، ونقطةُ تسجيلٍ واحدة
+
+**السياق.** الأدواتُ الأربع والثلاثون صار لها مضيفان: الإضافةُ المكتبية التي تقرأ نسخةَ الشاملة على جهاز المستخدم، وخادمٌ عامّ يقدّم الأدواتِ نفسَها عبر الشبكة. وكانت `sql.js` مثبَّتةً في صميم الشفرة — `catalog.ts` و`pages.ts` و`services.ts` تستدعيها مباشرة — ومعها استيرادُ `sql-wasm.wasm` الذي لا يفكّه إلا `esbuild`. فمن أراد قارئَ SQLite آخر لزمه أن يحمل معه بناءَ SQLite بلغة WebAssembly ولا يستعمله، ومن أراد اختبارَ الشفرة لزمه أن يزوّر استيرادَ الـ`wasm` (وهو ما كانت تفعله إضافةٌ في `vitest.config.ts`).
+
+**القرار.** فُصل ما يختلف بين المضيفَين عمّا لا يختلف:
+
+- `db.ts` يعرّف `ShamelaDb`: فتحُ ملفٍ للقراءة فقط، ومؤشِّرٌ عليه (`prepare/bind/step/get/reset/free`). لا شيء فيه من `sql.js`. و`sqljs.ts` هو تحقيقُها الذي تشحنه الإضافة، ويأخذ بايتات الـ`wasm` ولا يقرّر من أين جاءت.
+- `backend.ts` يجمع ما يختلف في `ShamelaDeps` وحده: أين النسخة (`resolvePaths`)، وكيف تُقرأ قواعدها (`db`)، وكيف يُبلَغ محرّكُ البحث (`createHelper`).
+- `register.ts` يصدّر `registerAllTools(server, deps)`: يسجّل الأدواتِ كلَّها على خادمٍ يملكه المستدعي، ويعيد إليه دورةَ حياة الخلفية ليسخّنها ويغلقها.
+- و`entry.ts` هو أوّلُ مستهلكٍ لهذه الواجهة، لا مسارٌ ثانٍ بجوارها: فيه — وفيه وحده — استيرادُ الـ`wasm`، وتوليدُ `sql.js`، وإطلاقُ الجافا، وناقلُ `stdio`.
+
+**ثمنه.** `Helper` صار واجهةً بعد أن كان صنفًا، والصنفُ الذي يشغّل الجافا اسمه الآن `JavaHelper`. وصار للمستودع بناءان من شجرةٍ واحدة: `dist/index.js` من `esbuild` للحزمة `.mcpb` كما كان، و`dist/lib/` من `tsc` ليُستورد. وحُذفت إضافةُ تزوير الـ`wasm` من `vitest.config.ts` لأنه لم يبقَ في الشفرة المشتركة استيرادٌ يزوَّر — وهذا هو المقياس الذي يقول إن الفصل تمّ: ما عاد أحدٌ يحتاج إلى حيلةٍ ليقرأ صفًّا من `master.db`.
+
+---
+
 ## ١١) المسحُ يسأل عن مجموعاتٍ متجاورة، لا عن كلماتٍ متناثرة — وكلُّ سؤالٍ خمسُ كلمات
 
 **السياق.** أداة تحرير محل النزاع (البند ٩٠) تسأل: أين كُتبت «لا خلاف» قرب هذه المسألة؟ وكان الظاهر أن `search_phrase` بوضع `near` تكفيها.
