+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "shamela-mcpb",
+  "name": "shamela-mcp",
   "version": "1.3.0",
   "description": "MCPB (MCP Bundle) for searching and studying a local Maktabah al-Shamela 4 install from Claude Desktop. 34 read-only tools.",
   "author": {
@@ -18,12 +18,24 @@
   "license": "MIT",
   "private": true,
   "type": "module",
-  "main": "dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/lib/server/index.d.ts",
+      "import": "./dist/lib/server/index.js"
+    },
+    "./sqljs": {
+      "types": "./dist/lib/server/sqljs.d.ts",
+      "import": "./dist/lib/server/sqljs.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "main": "dist/lib/server/index.js",
   "engines": {
     "node": ">=20"
   },
   "scripts": {
-    "build:server": "esbuild src/server/index.ts --bundle --platform=node --target=node20 --format=esm --loader:.wasm=binary --outfile=dist/index.js --banner:js=\"import { createRequire as __cr } from 'module'; import { fileURLToPath as __furl } from 'url'; import { dirname as __dn } from 'path'; const require = __cr(import.meta.url); const __filename = __furl(import.meta.url); const __dirname = __dn(__filename);\"",
+    "build:server": "esbuild src/server/entry.ts --bundle --platform=node --target=node20 --format=esm --loader:.wasm=binary --outfile=dist/index.js --banner:js=\"import { createRequire as __cr } from 'module'; import { fileURLToPath as __furl } from 'url'; import { dirname as __dn } from 'path'; const require = __cr(import.meta.url); const __filename = __furl(import.meta.url); const __dirname = __dn(__filename);\"",
+    "build:lib": "tsc -p tsconfig.lib.json",
     "build:java": "node scripts/build-java.mjs",
     "build": "npm run build:server && npm run build:java",
     "smoke": "tsx tests/smoke.ts",
@@ -38,20 +50,29 @@
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
     "test:unit": "vitest run tests/unit",
-    "test:integration": "vitest run tests/integration"
+    "test:integration": "vitest run tests/integration",
+    "prepare": "npm run build:lib"
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.29.0",
-    "sql.js": "^1.14.1",
-    "zod": "^3.24.1"
+    "sql.js": "^1.14.1"
   },
   "devDependencies": {
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/node": "^20.19.0",
     "@types/sql.js": "^1.4.11",
     "@vitest/coverage-v8": "^4.1.5",
     "esbuild": "^0.28.1",
     "tsx": "^4.19.2",
     "typescript": "^5.7.2",
-    "vitest": "^4.1.5"
+    "vitest": "^4.1.5",
+    "zod": "^3.25.76"
+  },
+  "types": "dist/lib/server/index.d.ts",
+  "files": [
+    "dist/lib"
+  ],
+  "peerDependencies": {
+    "@modelcontextprotocol/sdk": "^1.29.0",
+    "zod": "^3.25.76"
   }
 }
