+++ tests/fixtures/shared.ts
@@ -12,10 +12,12 @@ import * as path from "node:path";
 
 import { AyaIndexStore } from "../../src/server/ayaIndex/store.js";
 import { Catalog } from "../../src/server/catalog.js";
-import { Helper } from "../../src/server/helper.js";
+import type { ShamelaDb } from "../../src/server/db.js";
+import { JavaHelper } from "../../src/server/helper.js";
 import { PageStore } from "../../src/server/pages.js";
 import { resolveAll, type ShamelaPaths } from "../../src/server/paths.js";
 import { ServiceStore } from "../../src/server/services.js";
+import { createSqlJsDb } from "../../src/server/sqljs.js";
 import type { Backend } from "../../src/server/index.js";
 
 const requireFromHere = createRequire(import.meta.url);
@@ -25,7 +27,8 @@ let cachedPaths: ShamelaPaths | null = null;
 let cachedCatalog: Catalog | null = null;
 let cachedPageStore: PageStore | null = null;
 let cachedServiceStore: ServiceStore | null = null;
-let cachedHelper: Helper | null = null;
+let cachedHelper: JavaHelper | null = null;
+let cachedDb: ShamelaDb | null = null;
 let cachedBackend: Backend | null = null;
 
 /** Load sql.js wasm binary once (used by Catalog/PageStore/ServiceStore). */
@@ -34,6 +37,15 @@ export function getSqlWasm(): Uint8Array {
     const wasmPath = requireFromHere.resolve("sql.js/dist/sql-wasm.wasm");
     cachedWasm = new Uint8Array(fs.readFileSync(wasmPath));
     return cachedWasm;
+}
+
+/**
+ * The sql.js SQLite implementation, built once from the wasm read off disk —
+ * the same object the extension builds from the wasm esbuild inlined.
+ */
+export function getDb(): ShamelaDb {
+    if (!cachedDb) cachedDb = createSqlJsDb(getSqlWasm());
+    return cachedDb;
 }
 
 /** Resolve Shamela paths once. Throws ShamelaNotFoundError if not installed. */
@@ -47,8 +59,9 @@ export async function getPaths(): Promise<ShamelaPaths> {
 export async function getCatalog(): Promise<Catalog> {
     if (cachedCatalog) return cachedCatalog;
     const paths = await getPaths();
-    const wasm = getSqlWasm();
-    cachedCatalog = await Catalog.load(path.join(paths.database, "master.db"), wasm, { databaseRoot: paths.database });
+    cachedCatalog = await Catalog.load(path.join(paths.database, "master.db"), getDb(), {
+        databaseRoot: paths.database,
+    });
     return cachedCatalog;
 }
 
@@ -56,7 +69,7 @@ export async function getCatalog(): Promise<Catalog> {
 export async function getPageStore(): Promise<PageStore> {
     if (cachedPageStore) return cachedPageStore;
     const paths = await getPaths();
-    cachedPageStore = new PageStore(paths.database, getSqlWasm());
+    cachedPageStore = new PageStore(paths.database, getDb());
     return cachedPageStore;
 }
 
@@ -64,17 +77,18 @@ export async function getPageStore(): Promise<PageStore> {
 export async function getServiceStore(): Promise<ServiceStore> {
     if (cachedServiceStore) return cachedServiceStore;
     const paths = await getPaths();
-    cachedServiceStore = new ServiceStore(paths.database, getSqlWasm());
+    cachedServiceStore = new ServiceStore(paths.database, getDb());
     return cachedServiceStore;
 }
 
 /** Boot the Java helper once (5+ second JVM cold-start) and reuse across tests. */
-export async function getHelper(): Promise<Helper> {
+export async function getHelper(): Promise<JavaHelper> {
     if (cachedHelper) return cachedHelper;
     const paths = await getPaths();
-    cachedHelper = new Helper({ paths });
-    await cachedHelper.ready(30_000);
-    return cachedHelper;
+    const helper = new JavaHelper({ paths });
+    await helper.ready(30_000);
+    cachedHelper = helper;
+    return helper;
 }
 
 /** Build a Backend object that the MCP integration test can wire into createServer. */
