+++ vitest.config.ts
@@ -1,28 +1,6 @@
 import { defineConfig } from "vitest/config";
-import type { Plugin } from "vite";
-
-/**
- * Stub `.wasm` imports so vitest doesn't choke on
- * `import sqlWasm from "sql.js/dist/sql-wasm.wasm"` in src/server/index.ts.
- * Production esbuild bundles the wasm via `--loader:.wasm=binary`; tests
- * never touch this stub — they load the real wasm via fs.readFileSync in
- * tests/fixtures/shared.ts.
- */
-const wasmStubPlugin: Plugin = {
-    name: "wasm-stub",
-    enforce: "pre",
-    resolveId(id) {
-        if (id.endsWith(".wasm")) return "\0wasm-stub";
-        return null;
-    },
-    load(id) {
-        if (id === "\0wasm-stub") return "export default new Uint8Array(0);";
-        return null;
-    },
-};
 
 export default defineConfig({
-    plugins: [wasmStubPlugin],
     test: {
         globals: false,
         environment: "node",
@@ -40,7 +18,7 @@ export default defineConfig({
             provider: "v8",
             reporter: ["text", "html"],
             include: ["src/server/**/*.ts"],
-            exclude: ["src/server/index.ts"],
+            exclude: ["src/server/entry.ts"],
         },
     },
 });
